Make model render data frame a float to avoid loss of precision

+++ src/SharpLife.Game.Client/Entities/BaseEntity.cs
@@ -189,7 +189,7 @@ namespace SharpLife.Game.Client.Entities
                     RenderAmount = renderAmount,
                     RenderColor = RenderColor,
 
-                    Frame = (int)Frame
+                    Frame = Frame
                 };
 
                 modelRenderer.Render(ref renderData);
